+++ src/carl/core/Monomial_derivative.h
@@ -41,7 +41,7 @@ Term<Coefficient>* Monomial::derivative(Variable::Arg v) const
         {
             Monomial* m = new Monomial();
             m->mExponents.assign(mExponents.begin(), mExponents.end());
-            m->mExponents[(size_t)(it - mExponents.begin())].exp -= (size_t)1;
+            m->mExponents[(unsigned)(it - mExponents.begin())].exp -= (unsigned)1;
             m->mTotalDegree = mTotalDegree - 1;
             return new Term<Coefficient>(it->exp, m);
         }
